tests: Make command test structure common
Make the test command structure common so that multiple types of
statistics can be acquired from it.  Refactor the stack size test to
utilize the framework.
